Reorganize perception node code
The behavior and API provided remains the same, but the ROS parts have been decoupled from the image-diffing code. This might make it easier to provide alternative perception implementations if they fit in the same ROS interfaces.

The newly separated ROS wrapper no longer assumes that it has a node to itself. This allows it to be instantiated inside other Python scripts, like the pick state machine, which can simplify the debugging experience. Callers would still use the services, but they will resolve directly (I believe) to function calls in this case

Update camera name default
Change visualization to occur whenever a bin is captured
Print a warning if no images are being received