Integrate servo_to_pose capability in robot wrapper
Make robot wrapper class "controller aware"
Annotate methods with the controller they require. The newer @requires_controller(named) decorator ensures that the necessary controller is activated before the method gets invoked. This is a little overkill when we just need to swap back and forth between the traj_pos and raw group_pos controllers, but in the long term we'll have force control modes that we'll want to easily switch between and we'll benefit from having that management taken care of
Use servo to pose in MoveInOut state as first test