Track missing UOIS lib directory
The repo-wide gitignore has a (sensible) rule for ignoring
directories named lib, which are usually build outputs. However,
this original authors of this package used lib as the module
name for their manually-included 3rd party dependencies. We
need these files, so force adding them.
Add a simple gitignore in the data dir as documentation for the
data directory being kept empty