Track UOIS service in the main repo
This script has been living in a `demo_service` package
floating in workspaces.
This copy is also slightly modified to accept all images
over the ros service, instead of loading them directly from
disk. The segnet code still relies on them being stored to
disk (and it is the only code now that does so)